// frontend/src/enums/actions.ts
export enum Actions {
  INVERT_GAME_STATE = 'gameState/invert',
  TURN_INCREMENT = 'turn/increment',
  RESET_MOVES = 'move/reset',
  DECREMENT_MOVE = 'move/decrement',
  INCREMENT_MOVE = 'move/increment',
  ADD_MOVE = 'move/add',
  REMOVE_MOVE = 'move/remove',
  SET_NAME = 'name/set',
  ADD_RESULTS = 'results/add',
  SET_TREASURES_FOUND = 'treasures/set',
  RESET = 'reset',
  SET_HIGHSCORES = 'highscores/set',
}